+++ lottery-data/src/main/java/de/alpharogroup/lottery/wincategories/WinEvaluation.java
@@ -1,0 +1,19 @@
+package de.alpharogroup.lottery.wincategories;
+
+import de.alpharogroup.lottery.ticket.LotteryTicket;
+import lombok.*;
+import lombok.experimental.FieldDefaults;
+
+/**
+ * The class {@link WinEvaluation} represents a played lottery ticket.
+ */
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder(toBuilder = true)
+@FieldDefaults(level = AccessLevel.PRIVATE)
+public class WinEvaluation<C extends WinningOpportunity>
+{
+	C winCategory;
+	double evaluationOfUnitsResult;
+}