+++ lottery-data/src/main/java/de/alpharogroup/lottery/wincategories/WinningOpportunity.java
@@ -1,0 +1,8 @@
+package de.alpharogroup.lottery.wincategories;
+
+/**
+ * The interface {@link WinningOpportunity} is a marker for winning category classes
+ */
+public interface WinningOpportunity
+{
+}